robot_ws/src/omnibot_langchain/requirements.txt
anthropic>=0.28.0
langchain>=0.2.0
langchain-anthropic>=0.1.20
langchain-community>=0.2.0
requests>=2.31.0
pillow>=10.0.0
numpy<2.0
pyyaml>=6.0
